move exceptions module in the root folder
