+++ skore/src/skore/exceptions.py
